Move TourProvider to ExplorationAndAnalysis component so it's available when exported
